Refactor codebase to flatten structure and remove redundancy
- Merged `src/semantic/assembly_model.rs` into `src/semantic/assembler.rs`.
- Merged `src/parser/numerals.rs` and `src/parser/recursion.rs` into `src/parser.rs`.
- Merged `src/errors/assembly.rs` and `src/errors/messages.rs` into `src/errors.rs` (removing `src/errors/` dir).
- Merged `src/morphology/matcher.rs` into `src/morphology/mod.rs`.
- Moved regression test for dropped operator to unit tests in `src/semantic/expressions.rs`.
- Updated imports and verified tests pass.
