+++ tests/havoc.rs
@@ -1,4 +1,4 @@
-use glossa::parser::numerals::parse_greek_numeral;
+use glossa::parser::parse_greek_numeral;
 use glossa::parser::parse;
 use proptest::prelude::*;
